Create swarm defined type
This creates a lightweight swarm defined type, with a couple of scripts
which are assigned to either a manager or a working depending on the
assigned role.

To create a cluster, you must deploy at least a single manager, log into
the node, and run "sudo swarm_create_cluster". You will be presented
with both manager and worker tokens if successfully completed, and these
now need to be added to hieradata and Puppet redeployed. At least two
further managers should be joined, and then as many worker nodes as
required.

Using Terraform assign static IPs for the managers. As mentioned
previously, managers are snowflakes, and if they disappear then they
take cluster state with them. We should always provision at least 3
managers in any environment that matters, and so it makes sense to
provide static IPs so that we can hardcode the list of managers to check
when attempting to join nodes to the cluster.

New node types:

 - docker_base: for anything that should be included on a Docker Swarm machine
 - worker: used for creating a worker (ie a node that runs containers,
 but does not organisation)
 - manager: also runs containers within the cluster, but also does
 management