change nginx logging field names
This changes the names of fields in the nginx JSON log.  This is based
on the manual in the GDS Way[1].  The broad principles for naming are:

  1. follow the elastic beats naming conventions[2]
    - in particular, "sec" not "s" or "seconds"
  2. mark numerical data with a unit where appropriate (eg "bytes",
    "sec")
  3. be consistent with the filebeat nginx module names[3]
    - for example, I've preferred `ip` instead of `addr` in names like
      "remote_ip"

Note that $upstream_response_time is *not* marked with "sec" because
it's a list of times (one per upstream), not a single number.

Also note I've added a new field `access.url`; previously the URL was
available in the `request` field but that had some other cruft within
the same string (ie http method, version).

I'm not sure exactly how we decide what goes in the "access" object
and what stays outside, so I've put everything inside.

Due to whitespace changes, this commit is best viewed with the
--ignore-space-change option to the `git show` or `git diff` commands,
or by adding the ?w=1 query string to a github URL[4].

[1] https://gds-way.cloudapps.digital/manuals/logging.html#naming-conventions
[2] https://www.elastic.co/guide/en/beats/libbeat/current/event-conventions.html
[3] https://www.elastic.co/guide/en/beats/filebeat/current/exported-fields-nginx.html
[4] https://github.com/blog/967-github-secrets