Create cluster app defined type
My thinking in this commit is that we can use Puppet to manage:

 - monitoring
 - metrics
 - logging
 - environment variables
 - Docker Swarm creation and management

Ensure all configuration for apps is in hieradata so it's easier to
understand the configuration of an application.

We then use deployment scripts to manage the actual application
deployment, which is more akin to managing things in the PaaS.

Things to manage in deployment scripts may include:

 - Initial deployment of application
 - Forcing reloads after updates to env vars deployed in Puppet
 - Scaling up and down the number of replicas

We may want to change the way that we deploy secrets using the Docker
mounting of secrets, but this is just for testing purposes for the
time-being.