Enable email-alert-api-callbacks proxy
This adds a hostname, email-alert-api-callbacks, that is exposed
publicly which is used as a means for external services to route
requests in to email-alert-api.

It proxies the endpoint status-updates and will accept any subroutes
that start with that segment. Any other requests will receive a 404.

This is currently only enabled on integration until we have completed
authentication work.