Add OAuth credentials for authenticating-proxy in development VM
The authenticating-proxy app is needed to protect the draft stack. Other
apps which need OAuth credentials have them added in this way for the
development VM which means it's possible to use the Signon app's
make_oauth_work_in_dev script to update the apps in Signon with the
values from the OAUTH_ID & OAUTH_SECRET env vars for each app. Thus it
makes sense to follow the same pattern for the authenticating-proxy app.