SPIKE: Use authenticating-proxy to protect asset-manager
This introduces a new assets-draft virtual host which proxies requests
to the authenticating-proxy app which in turn now proxies requests to
the asset-manager app. Note that the setting of the X-Forwarded-Host
header seems to be important to make Signon authentication work, but it
may not be the best way to accomplish this.

This relies on the redirect_uri for the authenticating-proxy being set
to http://draft-assets.dev.gov.uk/auth/gds/callback in the Signon app so
that the original URL host matches that in the redirect_uri.

It also relies on the Asset Manager app redirecting requests for draft
assets on assets-origin.dev.gov.uk being redirected to
draft-assets.dev.gov.uk and on it checking the
X-Govuk-Authenticated-User header value which should be set by the
authenticating-proxy app.

Note that re-purposes the authenticating-proxy app which is normally
used to protect government-frontend. If we wanted to implement this for
real, we'd need to find a way to reinstate that behaviour in addition to
adding this new behaviour, either by adding a new instance of the
authenticating-proxy app or by making the authenticating-proxy app
capable of switching between two backends depending on the host in the
original URL.

Note also that we'd almost certainly want to use a higher-level
abstraction than nginx::config::site, we'd want it to be present in all
environments, and it might need to be on a different machine to the
authenticating-proxy - thus proxying to a port on localhost might not
cut the mustard.