SPIKE: Proxy assets-origin OAuth requests to asset-manager
This relies on a few things:

* The asset-manager app's redirect_uri being set to
http://assets-origin.dev.gov.uk/auth/gds/callback in Signon
* The user must have sign-in permission for asset-manager
* The asset-manager app needs to call authenticate_user! if the asset is
draft

Note that the exact form of the location block in the assets-origin
vhost is open to question, but this worked for me. I think the GDS SSO
functionality is sensitive to the host in the original request matching
the host in the OAuth redirect_uri. I think setting the X-Forwarded-Host
here means that they do match, but there might be a better way of doing
it.

Note that we might want to consider proxying to asset-manager via static
like we do for the other paths if that's really appropriate.