Set X-* headers set when proxying asset requests
Ensure the `X-Forwarded-Host` header is set when proxying asset requests
to static and other apps.  This header (along with `X-Real-IP`,
`X-Forwarded-Server` and `X-Forwarded-For`) is being set in the outer
server block but they were being hidden by the call to
`proxy_set_header` in these location blocks.

The `Host` header is set by default to match the host in the
`proxy_pass` directive so setting it manually is unnecessary.

See the nginx docs about `proxy_set_header`[1] for more information.

[1]: http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_set_header