Add draft-assets nginx vhost to development VM
We need this host to allow Asset Manager to protect draft assets behind
SSO. We're only adding the new virtual host to the development VM for
the moment, because we think we need to create SSL certificates, add DNS
entries, and create new ELB instances to make use of this in other
environments.

The configuration of the new virtual host is based heavily on that of
the assets-origin virtual host. One significant difference is that it
proxies all requests directly to Asset Manager - rather than doing this
via the static virtual host. Note that it is important that the
SSO-related paths (i.e. /auth/gds/*) as well as the Mainstream asset
paths (i.e. /mediai/*) and Whitehall asset paths (i.e.
/government/uploads/*).

We have intentionally not added a `proxy_set_header` directive to set
the `Host` request header inside the `location` block, because this
would mean that none of the `X-*` request headers set in the outer
`server` block would be set. In particular Asset Manager will need to be
able to rely on the value of `X-Forwarded-Host` being set correctly in
order to determine whether an asset request came via `assets-origin` or
`draft-assets`.

The new virtual host can be enabled in other environments by adding the
following snippet to the body of `modules/router/manifests/nginx.pp`.

    class { 'router::draft_assets':
      real_ip_header => $real_ip_header,
    }

We have done some preliminary testing of the above snippet in the
integration environment and it seemed to work correctly.