Add a jenkins job to extract data from graphite
See https://github.com/alphagov/app-performance-summary/pull/1

I've installed pipenv as a way to manage dependencies. This
is the preferred python packaging tool now and makes it easier
to use modern versions of python as encouraged by the GDS way.
See https://docs.pipenv.org/

I've disabled proper version pinning because the python installation
on these machines is a bit of a mess. Python 3.4 is available through
the distribution but it's running a very old version of pip that
is missing some required functionality. I'm not hugely concerned
about this for now.